+++ Readme.md
@@ -14,6 +14,7 @@ I [built this for the web in Vue.js](https://github.com/Plsr/DoDee) a few years 
 - [ ] Design
 - [x] Allow todos to be checked off
 - [ ] Add settings page
+- [ ] Add splash screen
 - [ ] Delete Todos older than two days
 - [ ] Display todos from the last day for import
 - [ ] Allow todos to be edited
